refactor: CodeEditor 컴포넌트 리팩토링 fix: 탭 전환시 브레이크포인트 수정아 안되던 문제 수정
* 에디터 설정 파일을 constants/monaco/editor-config.ts로 분리

* 브레이크포인트 로직을 hooks/editor/useBreakpointManager.ts로 분리
